+++ tp1/MySphere.js
@@ -6,7 +6,7 @@
     CGFobject.call(this,scene);
    	var values = args.split(" ");
 	for(var i = 0; i < values.length; i++){
-		values[i] = parseInt(values[i]);
+		values[i] = parseFloat(values[i]);
 	}
 
 	this.radius = values[0];
